+++ Cargo.toml
@@ -12,23 +12,26 @@ repository = "https://github.com/wicrs/server"
 [dependencies]
 base64 = "0.13"
 oauth2 = "3.0"
-rand = "0.7"
-reqwest = { version = "0.10", features = ["json"] }
+rand = "0.8"
+reqwest = { version = "0.11", features = ["json"] }
 serde = "1.0"
 serde_json = "1.0"
-bincode = "1.3.2"
-tantivy = "0.14.0"
+bincode = "1.3"
+tantivy = "0.14"
 sha3 = "0.9"
-tokio = { version = "1.5", default-features = false, features = ["macros", "fs", "time", "rt", "net", "sync"] }
+tokio = { version = "1", default-features = false, features = ["macros", "fs", "time", "rt", "net", "sync"] }
 futures-util = { version = "0.3", default-features = false, features = ["async-await", "sink", "std"] }
 uuid = { version = "0.8", features = ["serde", "v4"] }
-actix = "0.10"
+actix = "0.11"
 actix-web = "3"
 actix-web-actors = "3"
 tokio-tungstenite = "0.14"
-lazy_static = "1.4.0"
+lazy_static = "1.4"
 futures = "0.3"
 parse-display = "0.4"
+warp = "0.3"
+async-graphql = { version = "2.0", features = ["uuid", "string_number"] }
+async-graphql-warp = "2.0"
 
 [profile.release]
 lto = true
