+++ src/main.rs
@@ -17,7 +17,7 @@ fn load_config(path: &str) -> wicrs_server::config::Config {
 
 /// Main function, loads config and starts a server for the HTTP API.
 #[actix_web::main]
-async fn main() -> std::io::Result<()> {
+async fn main() {
     let config = load_config("config.json");
-    wicrs_server::httpapi::server(config).await
+    wicrs_server::httpapi::graphql(config).await;
 }
